Rename store package to upload
