+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptions.java
@@ -12,16 +12,16 @@ import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.commandline.Validator;
 import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocatingTransformer;
 import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
-import com.teamscale.jacoco.agent.store.IUploader;
-import com.teamscale.jacoco.agent.store.NoopUploader;
-import com.teamscale.jacoco.agent.store.UploaderException;
-import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageConfig;
-import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageUploadStore;
-import com.teamscale.jacoco.agent.store.upload.delay.DelayedCommitDescriptorStore;
-import com.teamscale.jacoco.agent.store.upload.http.HttpUploadStore;
-import com.teamscale.jacoco.agent.store.upload.teamscale.TeamscaleUploader;
 import com.teamscale.jacoco.agent.testimpact.TestExecutionWriter;
 import com.teamscale.jacoco.agent.testimpact.TestwiseCoverageAgent;
+import com.teamscale.jacoco.agent.upload.IUploader;
+import com.teamscale.jacoco.agent.upload.NoopUploader;
+import com.teamscale.jacoco.agent.upload.UploaderException;
+import com.teamscale.jacoco.agent.upload.azure.AzureFileStorageConfig;
+import com.teamscale.jacoco.agent.upload.azure.AzureFileStorageUploadStore;
+import com.teamscale.jacoco.agent.upload.delay.DelayedCommitDescriptorStore;
+import com.teamscale.jacoco.agent.upload.http.HttpUploadStore;
+import com.teamscale.jacoco.agent.upload.teamscale.TeamscaleUploader;
 import com.teamscale.jacoco.agent.util.AgentUtils;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.EDuplicateClassFileBehavior;
