+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/IUploader.java
@@ -1,4 +1,4 @@
-package com.teamscale.jacoco.agent.store;
+package com.teamscale.jacoco.agent.upload;
 
 import java.io.File;
 
@@ -8,7 +8,7 @@ public interface IUploader {
 	/** Uploads the given coverage file. */
 	void upload(File coverageFile);
 
-	/** Human-readable description of the store. */
+	/** Human-readable description of the uploader. */
 	String describe();
 
 }
