+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/UploaderException.java
@@ -1,4 +1,4 @@
-package com.teamscale.jacoco.agent.store;
+package com.teamscale.jacoco.agent.upload;
 
 import okhttp3.ResponseBody;
 import retrofit2.Response;
@@ -6,7 +6,7 @@ import retrofit2.Response;
 import java.io.IOException;
 
 /**
- * Exception thrown from an upload store. Either during the upload or in the validation process.
+ * Exception thrown from an uploader. Either during the upload or in the validation process.
  */
 public class UploaderException extends Exception {
