+++ agent/src/test/resources/com/teamscale/jacoco/agent/d/upload/delay/git-properties.jar
